requirements.txt
flask
torch 
transformers
python-dotenv
flask-login 
requests-oauthlib